Initial white-label config for SV Customer access
This branch should be deployable as a customer-accessible version of
MonsterUI with locked-down functionality (preventing them from spending
money on our behalf, as well as trying to prevent them from doing too
much damage to themselves).

This commit modifies the MonsterUI built-in config options to remove DID
purchasing and porting capability, hides the App Store, and prevents new
account creation.